add bin/stage.py tool
Added a new `bin/stage.py` tool to move, with proper locking,
the submit file under the staged directory, and to update
slot's JSON file to indicate that the submit file has been collected.

The `bin/stage.py` tool will print the SHA256 hash of the submit
file that was moved under the staged directory along with the
count of the files moved under the unexpected directory is all is well.
If the `bin/stage.py` tool detects an error, it will print
"exit.code", where "code" is the exit number, along with the
count of the files moved under the unexpected directory.

Improved or fixed a number of debug messages.

Improved function call tree for when `stage_submit()` is called.

Changed `stage_submit()` to return two values, hexdigest and count.

Added `prerr()` to `iocccsubmit/ioccc_common.py` to print to stderr.

Changed python package in `Makefile` from 0.11.2
to 0.11.3.

Changed `iocccsubmit/ioccc_common.py` VERSION_IOCCC_COMMON from "2.5.2 2025-02-04"
to "2.5.3 2025-02-03".

Tested the above with `make clobber all install` on macOS.

TODO: Test setup on submit.ioccc.org.