release sbin/collect.sh
Added extensive error checking and bug fixes to `sbin/collect.sh`.
Performed extensive testing under a number of ways that slots
and files could be corrupted.

Problems discovered by `sbin/collect.sh` are reflected back to the slot
status message so that the submitter may re-submit.

When `sbin/collect.sh` discovers a problem, and the destination
is under git control, the problems are reflected in the git commit
message in detail.  If the destination is not under git, or otherwise
use of git is disable (for example, by using the `-G` flag), then
the problems are written to stderr.

We are now ready to automate collection of submit files via a cron job.

Changed `sbin/collect.sh` VERSION from "2.1.1 2025-02-06"
to "2.3.0 2025-02-08";

Tested the above with `make clobber all install` on macOS.

TODO: Setup a cron job to run this code and collect submissions
      in the area where the IOCCC judges do their work.